moving worker to cmd
