+++ cmd/worker/main.go
@@ -17,11 +17,12 @@ package main
 import (
 	"time"
 
-	"github.com/undistro/zora/worker/run"
 	"go.uber.org/zap/zapcore"
 
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/log/zap"
+
+	"github.com/undistro/zora/pkg/worker/run"
 )
 
 var log = ctrl.Log.WithName("worker")
